Common code was moved into a single reusable component
