+++ src/app/globals.css
@@ -10,7 +10,7 @@
 body{
   width: 100vw;
   min-height: 100vh;
-  background-image: linear-gradient(rgba(9, 0, 77, 0.65), rgba(9, 0, 77, 0.65)), url(assets/images/background-1.jpg);
+  background-image: url(assets/images/background-1.jpg);
   background-size: cover;
   background-position: center;
   padding: 10px, 8%;
