--- src/app/offers-photographers/page.module.css
@@ -1,241 +1,0 @@
-/* @import '~react-datetime/css/react-datetime.css'; */
-
-.artist_description_tiles{
-    display: flex;
-    align-items: center;
-    justify-content: space-around;
-    font-size: 22px;
-    color: #fff;
-    margin: 0px;
-    border-radius: 10px;
-    box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
-    padding: 10px 7px 4px 7px;
-    width: 790px;
-    height: 282px;
-    margin-left: auto;
-    margin-right: auto;
-    background-color: #775a967c;
-    margin-top: 20px;
-}
-
-.artist_description{
-    margin-left: 10px;
-    text-align: justify;
-}
-
-.artist_description_tiles img{
-    height: 270px;
-    width: 270px;
-    border-radius: 5px;
-}
-
-.artist_description_tiles p{
-    text-align: left;
-    line-height: 1.3;
-    font-size: 15px;
-    margin-top: 10px;
-    width: 500px;
-}
-
-.artist_description_tiles .provided-services{
-    font-size: 20px;
-    width: 900px;
-    margin-left: 20px;
-}
-
-.artist_description_tiles button{
-    align-items: right;
-}
-
-.artist_description_tiles .artist-description h3{
-    margin: 20px 0px;
-}
-
-.stars img{
-    width: 24px;
-    height: 22px;
-    margin: 10px 5px 0px 0px;
-}
-
-.service_tiles:hover,
-.artist_tiles:hover{
-    transform: translateY(4px);
-    background-color: rgba(113, 97, 128, 0.65);
-}
-
-.tile_clickable{
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    flex-wrap: wrap;
-    margin-top: 10px;
-    margin-left: auto;
-    margin-right: auto;
-    background-color: none;
-}
-
-.tile_clickable a{
-    text-decoration: none;
-}
-
-.tile_clickable p{
-    color: #fff;
-}
-
-.tile_clickable img{
-    margin-bottom: 10px;
-}
-
-.service_tiles{
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    margin: 8px;
-    border-radius: 10px;
-    background-color: #775a96c3;
-    box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
-    margin-top: 10px;
-    padding: 5% 15% 5px;
-    width: 186px;
-    height: 290px;
-    color:#fff;
-    line-height: 1;
-    text-align: center;
-}
-
-.service_tiles h2{
-    color: #fff;
-    font-size: 25px;
-    padding: 50px 0;
-    text-align: center;
-}
-
-.title h2{
-    text-transform: uppercase;
-}
-
-.service_tiles p{
-    font-size: 20px;
-    width: 100px;
-    margin: 10px 0px 0px 0px;
-    text-transform: uppercase;
-}
-
-.service_tiles h3{
-    width: 100px;
-    font-size: 20px;
-    padding-bottom: 0px;
-}
-
-.service_tiles ul{
-    font-size: 13px;
-    text-align: left;
-    line-height: 1.4;
-}
-
-.service_tiles button{
-    margin-left: 0px;
-}
-
-.service_tiles button{
-    width: 90px;
-    padding: 10px 0px;
-    background: transparent;
-    color: #fff;
-    font-size: 12px;
-    border: 1px solid #d29de8d3;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: 0.5s;
-}
-
-.service_tiles button:hover{
-    transform: translateY(-5px);
-    background:#d29de8d3;
-    color: #121414;
-}
-
-.btn_timeslot_weekday,
-.btn_timeslot_time{
-    margin: 8px;
-    border-radius: 50px;
-    background: #775a967c;
-    border: #775a967c;;
-    box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
-    color: #fff;
-    font-size: 14px;
-    cursor: pointer;
-    transition: 0.3s;
-}
-
-.btn_timeslot_weekday{
-    width: 60px;
-    height: 70px;
-}
-
-.btn_timeslot_time{
-    width: 100px;
-    height: 40px;
-}
-
-.btn_timeslot_weekday:hover,
-.btn_timeslot_time:hover{
-    transform: translateY(-5px);
-    background:#d29de8d3;
-    color: #121414;
-}
-
-.btn_go_to_shopping_cart{
-    display: block;
-    width: auto;
-    margin: auto;
-    padding: 10px 20px;
-    margin-top: 40px;
-    margin-bottom: 40px;
-    height: 50px;
-    background: transparent;
-    color: #fff;
-    font-size: 20px;
-    border: 1px solid #d29de8d3;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: 0.5s;
-}
-
-.btn_go_to_shopping_cart:hover{
-    transform: translateY(-5px);
-    background:#d29de8d3;
-    color: #121414;
-}
-
-.btn_timeslot_time input{
-    margin: 0.8rem 0.3rem 0.8rem 0.8rem;
-}
-
-.btn_timeslot_weekday input{
-    margin: 0.8rem 0.3rem 1rem 0.8rem;
-}
-
-.date_time_picker{
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    flex-wrap: wrap;
-    margin-top: 10px;
-    margin-left: auto;
-    margin-right: auto;
-    background-color: none;
-
-    background: transparent;
-    color: #fff;
-    font-size: 1.3rem;
-    border: .1rem solid #d29de8d3;
-    border-radius: .3rem;
-    width: 15rem;
-}
-
-.btn_go_to_shopping_cart:disabled{
-    background-color: grey;
-    cursor: not-allowed;
-    transform: translateY(0px); 
-}
